requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10
APScheduler==3.10.4
netifaces==0.11.0
ping3==4.0.4
Pillow==10.0.0
flask-sqlalchemy==3.0.5
opencv-python==4.8.1.78